Fix BAxUS expansion test for bool tensor argmax

+++ tests/optim/test_baxus_strategy.py
@@ -86,14 +86,19 @@ def test_expansion_splits_every_splittable_parent_bin() -> None:
         seed=8,
         state=state,
     )
-    parent_assignment = strategy.embedding.ne(0).argmax(dim=1)
+    parent_assignment = strategy.embedding.ne(0).to(torch.int64).argmax(dim=1)
 
     assert strategy.expand_subspace()
     assert strategy.target_dim == 6
     assert torch.all(strategy.embedding.ne(0).sum(dim=1) == 1)
     assert torch.all(strategy.embedding.ne(0).sum(dim=0) > 0)
     for parent in range(2):
-        child_bins = strategy.embedding[parent_assignment == parent].ne(0).argmax(dim=1)
+        child_bins = (
+            strategy.embedding[parent_assignment == parent]
+            .ne(0)
+            .to(torch.int64)
+            .argmax(dim=1)
+        )
         assert torch.unique(child_bins).numel() == 3
